=typ load typed extensions as part of a separate untyped extension
This is needed because the `ActorSystemAdapter` is instantiated early on
in the typed ActorSystem initialization process already when the infrastructure
is created to run the typed guardian. This is too early to initialize
the typed extensions. The new untyped LoadTypedExtension is only loaded
later on when the untyped ActorSystem is ready.