=htc #380 TLS should close connection when instructed by stream completion
Before introduction of the `delayCancellation` stage TLS closing worked
despite of the default `TLSCLosing.ignoreComplete` mode somewhat
accidentally because cancellation would close the TCP connection.

With the introduction of `delayCancellation`, stream completion is the signal
to act on.

Without this fix, the server TLS side does not close the connection
eagerly but only the idle-timeout will do so later on.