Fix issue with dead actors in pubsub topics
We have run into out-of-memory problems and it seems like the subscribers list is never emptied in the pubsub topic class. By looking at the code it seems like the remove code indeed isn't run. Here's a guess on how it could be fixed, there's a lot of details here that I don't know anything about so please correct me if I'm wrong.