Allow `groupBy` to recreate already closed substreams #24758
`groupBy`'s default behavior is to filter elements, that would go to a substream
that is already closed. This allows `groupBy` to recreate those already closed streams,
if wanted, and run them as if they would've never run before in the first place.

Overload instead of default parameter

Make set of closed keys unmodifiable if not needed

Adjust documentation, adding a warning on memory consumption.

Add MiMa exclude.