Allow named pipes / FIFOs when using `FileIO` (fixes #25328).
`FileIO` was using NIO's `isRegularFile` to check whether the path
passed was a file and not a directory (added in #21801) - however this
has the side-effect of also excluding things that are not considered
regular files in POSIX terminology, like named pipes / FIFOs.

The requirement has been swapped to only check if something is not a
directory, allowing anything else - this could mean a user can try to
open other non-regular files like devices, but that check is probably
not necessary.