Set Terminated.failure when guardian actor terminates with a failure Refs #25345
Wrap the guardian actor in a behavior that completes a promise when that
actor terminates. This behavior wrapper records exceptions that
terminate the actor too.

The future generated from this promise is passed into the ActorSystem
using a specialised Setup. This is required as there is no easy way to
pass values from the method which creates the actor and the untyped
ActorSystem to ActorSystemAdaptor where whenTerminated is
implemented.