Allow OptionVal to hold boxed primitives
Previously OptionVal was restricted to types that have Null as their
bottom type, that is reference types (that extend AnyRef/Object), rather
than also allowing primitives (that extend AnyVal).

By removing this restrictions any type is a valid type argument.

The effect this will have is that primitives will be box, because we
need a way to represent the None case (using the null value).