Fix StreamRef termination on node failure
StreamRefs handle failure of the remote side of the conversation if the
remote actor gets terminated, but in the case of full node failure
(e.g. node killed or downed after network partition), the StreamRef
doesn't receive a termination message from its partner. This is because the
FunctionRef actor type is used which doesn't extend the DeathWatch
trait. FunctionRef receives Terminations from the stopped actor, but it
doesn't listen for address down events, and the Remote Watcher doesn't
notify it, as it assumed it implemented DeathWatch.