!str remove {Source,Flow}WithContext.statefulMapConcat, fixes #26330
statefulMapConcat will by design confound processing of single elements.
However, in the WithContext APIs the user only has access to the elements
but not to the contexts so the association is likely to be wrong in the end.

So, for now, we follow our basic rule not to include potentially order-changing
operations in the WithContext APIs and remove statefulMapConcat. Users that
need something like that should drop to the regular tuple passing APIs and
manage context propagation manually.