clarifies backoff's "restart" #23406
The doc is using the verb "restart" whereas the supervisor
is actually stopping the actor and starting it
again. This might cause some confusion, e.g. if used
along with stash, user expects the actor to be restarted
because the messages are unstashed in the `preRestart` hook.

This commit avoids using the "restart" verb as much as possible as well
as adding a note regarding the use of a backoff with a stash.