Allow build to run on JDK 8 (#26642)
Ref #26233

Currently Akka fails to build on sbt community build, which runs on JDK 8. Likely it will fail on Scala community build as well due to `--ignore-source-errors`.
This makes the flag conditional based on the running JDK.

It will be useful for Scala toolchain to be able to validate its latest using latest Akka if possible.