Update MiMa: drop 2.4, add 2.5
We promise binary compatibility across minor versions, which would in theory
mean we should check compatibility between 2.4 and 2.6.

However, since 2.4 is EOL, we no longer guarantee bincompat with it. In
practice this should not make much of a difference, since only in rare cases
would a change be binary compatible with 2.5 but not with 2.4.