Don't branch while releasing
Don't introduce a new branch while releasing, so the release is just a tag on
master.

The approach with `version.sbt` is intended to be temporary: once we have a
release with a tag on master, we can introduce a plugin like sbt-dynver to
determine the version based on the tag instead of generating a `version.sbt`.
See also #26675.