Log rebalance progress at INFO per rebalance info
For cases where rebalance takes a long time due to a relatively large cluster change
e.g. when doubling number of nodes (common for 1-2, 3-6 etc) there is
little insignt into what is happening unless DEBUG logging is enabled.

This adds an INFO log per rebalance-interval (default is 10s) to show
what is in progress and which new shards are starting rebalancing.