Allow shard rebalancing to be based on a factor
Currently, the rebalance-threshold is used to decide how many shards
should be rebalanced per rebalance iteration, with
max-simultaneous-rebalance limiting the number of ongoing hand offs in
cases where it takes more than the rebalance interval to hand off a
shard. In some environments, this can make rebalancing too slow, for
example, with 100 shards, going from one node to two nodes takes 8
minutes to completely rebalance with the default configuration. The
alternative is to select a higher rebalance threshold, which will result
in less even shard allocations.

This separates the rebalance-threshold configuration from the number of
shards that gets rebalanced, and allows it to be an absolute number, or
a factor of the total number of shards that the region with the most
shards has.