Use SPDX short identifier for license (#27277)
* Use SPDX short identifier for license

The SPDX license list (https://spdx.org/licenses/) defines a standard
for identifying licenses that allows tooling to unambiguously identify a
license. This changes Akka to use the SPDX short identifier for the
Apache 2 license.

Any developer wishing to publish their own build of Akka to bintray will
need to make this change, as bintray uses the SPDX short identifier to
verify that artifacts being published to free accounts are indeed open
source. So this will make it easier for people to publish their own
builds.

* Update Apache license URL

Co-Authored-By: Arnout Engelen <github@bzzt.net>