Fixes intermittent "Could not create directory" issue
Fixes #27661

I saw that MultiJVM's assembly/fullClasspath contains a bunch of duplicates in the form of:

```
[info] * Attributed(/Users/eed3si9n/work/akka/akka-cluster-sharding/../akka-protobuf-v3/target/akka-protobuf-v3-assembly-2.6-SNAPSHOT.jar)
[info] * Attributed(/Users/eed3si9n/work/akka/akka-cluster/../akka-protobuf-v3/target/akka-protobuf-v3-assembly-2.6-SNAPSHOT.jar)
```

Helena pointed out to me that project/Protobuf.scala introduces `/../` in the middle of an unmaged JAR path:

```
    // this keeps intellij happy for files that use the shaded protobuf
    Compile / unmanagedJars ++= Seq(
        baseDirectory.value / ".." / "akka-protobuf-v3" / "target" / s"akka-protobuf-v3-assembly-${version.value}.jar"),
```

While that may keep IntelliJ, sbt-assembly is not very happy with the duplicated protobuf jars.

This works around the issue by calling `getCanonicalFile` to make the path canonical.