Report UDP connection error for negatively-cached hostnames (#28444)
When creating an UDP connection for a hostname for which we already have a
'negative cache' saying it cannot be resolved, report that as an error
as well.

Noticed while looking at #28133, but looking closer at the logs that seems
like a different problem. Added logging that might make it easier to diagnose.