Prevent infinity high CPU consumption loop
This commit contains two changes.

1. It introduced to use `Thread.yield()` as a cheap and easy way to ask JVM to
spent some time at another thread to prevent 100% usage of CPU by a thread
that's just checking the next.

2. Merged condition to enter and exit from the loop.

The second part commit has prevented a bug when `LightArrayRevolverScheduler`
going into an infinite loop inside `nextTick`.

This loop is infinity because the code has only one place where `tail` may have
`null` as the next element and this place inside`pollNode` that is in the loop
in the fact.

Probably it also might be achieved over `addNode` but in case if Akka system is
terminating or terminated, this thread will loop for infinity time.

This changes also prevent infinity loop when `next` is same as `tail.next()` and
it doesn't match `get()`.

This fixed an issue / symptopms that I've discribed at #28601