Make it obvious a custom serializer is needed
When I read the doc, I get the impression that if no custom serializer is implemented, akka will figure it out by leveraging `ReplicatedDataSerializer`, since in my use case, I do nothing but combining 2 `GCounter` to build a `FractionCounter`. However, my impression is wrong, and I have to read the source code to understand. Maybe better to make it obvious