Publish snapshots to gustav with 'publish'
This removes the need to publish to a 'predictable' directory and then
'manually' rsyncing the locally-created repository from jenkins to gustav.

This is not needed for anything, it just simplifies the build a bit. We
could eventually use it for the paradox and api docs as well, replacing
the current custom `DeployRsync`.