Extend ContextMapStrategy to mapAsyncUnordered
This extends the ContextMapStrategy idea to mapAsyncUnordered. This
might help motivate the feature:

Previously, a big limitation of FlowWithContext is that it was somewhat
unclear what kinds of operations were allowed on streams in order not to
violate the expectations of the sinks that use that context.

For example, the current approach for dealing with Kafka commits in
contexts would allow filtering, but not reordering of the elements.

By putting this information in the type of the ContextMapStrategy,
components that care about the offsets (like Alpakka Kafka) can provide
a strategy (or multiple) that allows only those operations that are safe
for them.

Components that provide context-aware flows but don't care about the
offsets themselves (such as generic alpakka components) can indicate
their behaviour in the type of the strategy they accept.

The test shows 2 possible Kafka strategies, one allowing iteration (but
not reordering) and one allowing reordering (but not filtering), and 2
example streams showing how this can be composed in a typesafe way.

This still doesn't cover flatMapConcat. For flatMapConcat, there are some
trade-offs: i.e. in the Kafka case, if you absolutely want all input elements
to be committed, AFAICS we would have to delay each element (until we know
whether the next event is the next element or the stream completion), which is
quite a cost. If you're OK with skipping some commits for filtered-out
input elements (which I'd say is typically reasonable), you don't have to
incur this cost. Signalling whether this is necessary could also be done
from the mapping strategy, and we could provide 'strict' and 'loose' strategies
corresponding the the different behaviours. That seems to fit rather nicely.

(I think there is another option that would only delay the commits, also fit
flatMapMerge and reordering (even at the same time), while only possibly missing
commits at the 'end' of the stream - but let's not get into the woods too far
yet ;))