Add apply methods for typed ActorTestKit allowing for passing an ActorSystem or ActorSystemSetup (#27338).
Added an apply method that takes a akka.actor.typed.ActorSystem and generates a ActorTestKit.
Added an apply method that takes a string and akka.actor.setup.ActorSystemSetup and generates an ActorTestKit

Modified the internal ActorTestKit class to use an ActorSystemSetup instead of a config and updated the other apply methods to generate an ActorSystemSetup in place of the config.
This was an attempt to maintain compatability.

added the related test specs for the new apply methods and updated the apply methods that changed.