Make Output invariant
Turns out Scala 3 is right.  Here's an example: let's say we have 'class
Animal' and 'class Cat extends Animal' if have an Output[Animal] with
Output[+T] that might be a Pull[Promise[Option[Cat]]]] which wouldn't
deal well with receiving some other animal than Cat.  So the cast is
unsafe.

Now ideally Output would be contravariant, because it's intended to be
the sink that consumes values of type T.  Then Cancel would just satisfy
any Output (Output[Any]).  However Promise's are invariant in their type
T, because it exposes itself as a Future[T] (which is covariant).

Btw, Scala 3 accepts:

    case object Cancel extends Output[_]

which I think would allow the use of a single Cancel instead of one per
type T.