stream: when coalesce-writes is enabled loop through stageActor
Before it was cycling to the TCP connection actor. Sending messages to
idle actors can be relatively expensive if a mailbox needs to be
scheduled and a pool thread needs to be woken up.

Looping through the stageActor means effectively scheduling an
asyncCallback to the shortcutCircuit which is quite cheap