Update scala-java8-compat to `1.0.0`
The compat library added an `early-semver` versionScheme which means
that 1.0.0 is marked as binary incompatible with any version lower
than 0.9.1 - even though it's not.

For example:

```
[error] (update) found version conflict(s) in library dependencies; some are suspected to be binary incompatible:
[error]
[error] 	* org.scala-lang.modules:scala-java8-compat_2.13:1.0.0 (early-semver) is selected over 0.9.0
[error] 	    +- com.github.blemale:scaffeine_2.13:4.1.0            (depends on 1.0.0)
[error] 	    +- com.typesafe.akka:akka-actor_2.13:2.6.14           (depends on 0.9.0)
```

The `scalaModuleMimaPreviousVersion` has been set to `0.9.1` in the project, so upgrading
Akka to this should allow it to work with a lot of other libraries who have already done
the major update (supports Scala 3) to 1.0.0.