Test the same scenario on Scala 2.13 and 3
A case object is `Serializable` on Scala 3 but not on Scala 2.13, so
this test would fail because no serializer is found on Scala 2.13,
and because the Java serializer is disabled on Scala 3.

By using a class instead, which is not `Serializable` both on Scala 2.13
and Scala 3, we test the same thing on both Scala versions.