Simplify usage of java8-compat and use 1.0.0.
The current behavior in Akka is that the >= 2.13 cross version gets the
1.0.0 version of java8-compat, although everything else still gets
0.8.0. This behavior was last updated in
https://github.com/akka/akka/commit/466802c641721447826da39f29c62a7c4bcecc09,
and originally added in
https://github.com/akka/akka/commit/49322afae0c7c0f96c76d5131207ac55a8f92cd3,
but I believe it can be simplified further to always use 1.0.0.
1.0.0 is cross published for 2.11, 2.12, 2.13, and even 3.0, so this
shouldn't really be an issue using it across the board.

The main reason for this change is that users that are still on 2.12 are
starting to get errors like:

```
[error]         * org.scala-lang.modules:scala-java8-compat_2.12:1.0.0 (early-semver) is selected over 0.8.0
[error]             +- com.github.swagger-akka-http:swagger-akka-http_2.12:2.5.0+13-a0fac563+20210920-1710-SNAPSHOT (depends on 1.0.0)
[error]             +- com.typesafe.akka:akka-actor_2.12:2.6.16           (depends on 0.8.0)

```

Since much of the ecosystem is bumping to the 1.0.0, but 2.12 of Akka is
still pulling in 0.8.0. This will avoid nasty hacks like:

```scala
ThisBuild / libraryDependencySchemes += "org.scala-lang.modules" %% "scala-java8-compat" % "always"
```