Correct outdated external reference to erlang.org
Section "Discussion: Why No Guaranteed Delivery?" mentions section 10.10 of the Erlang Academic FAQ "Academic and Historical Questions" that is currently titled as "10.10 What limits does Erlang have?". Section 10.10 of the FAQ has been changed and is not related to the topic anymore.

FAQ topics "10.8 Is the order of message reception guaranteed?" and "10.9 If I send a message, is it guaranteed to reach the receiver?" seem to match the original intended references.